Refactor metric helpers into its own file to deduplicate
